doi type date time to long

+++ backend/src/main/java/com/example/demo/entity/Category.java
@@ -18,7 +18,7 @@ public class Category {
     @GeneratedValue(strategy =  GenerationType.AUTO)
     private int idCategory;
 
-    private int createdTime;
+    private long createdTime;
 
     @NotNull
     private String nameCategory;
